wip add db helper func, caching for star network
